+++ packages/terra-graphs-docs/src/terra-dev-site/doc/example/Graphs-Carbon/PairedResultGraph/BasicPairedResultGraphs/TimeseriesWithDateline.jsx
@@ -1,0 +1,44 @@
+import React from 'react';
+import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
+import Carbon from '@cerner/carbon-graphs/lib/js/carbon';
+import '@cerner/terra-graphs-docs/lib/Css/ExampleGraphContainer.module.scss';
+import '@cerner/terra-graphs/lib/components/Graph.module.scss';
+import '@cerner/terra-graphs/lib/components/PairedResult/PairedResultGraph.module.scss'; import pairedResultTimeseries from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/PairedResult/pairedResultTimeseries';
+import exampleData from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/PairedResult/pairedResultTimeseriesData';
+import {
+  loadDatelinePopup,
+} from '@cerner/terra-graphs-docs/lib/example-datasets/popup';
+/*
+Please refer to the documentation below to see the graphConfig and data objects
+*/
+const graphConfig = utils.deepClone(pairedResultTimeseries('#pairedResultWithDateLine'));
+graphConfig.dateline = [
+  {
+    showDatelineIndicator: true,
+    label: {
+      display: 'Action Date',
+    },
+    color: '#C97318',
+    shape: Carbon.helpers.SHAPES.SQUARE,
+    onClick: loadDatelinePopup,
+    value: new Date(2017, 10, 1).toISOString(),
+  },
+];
+const dataset = utils.deepClone(exampleData);
+
+const PiaredResultExample = () => {
+  React.useEffect(() => {
+    const graph = Carbon.api.graph(graphConfig);
+    dataset.forEach((data) => {
+      graph.loadContent(Carbon.api.pairedResult(data));
+    });
+  }, []);
+  return (
+    <React.Fragment>
+      <div id="tooltip" className="initial-tooltip" />
+      <div id="pairedResultWithDateLine" />
+    </React.Fragment>
+  );
+};
+
+export default PiaredResultExample;