+++ packages/terra-graphs-docs/CHANGELOG.md
@@ -21,3 +21,5 @@
 
 * Changed 
   * Moved core documentation to terra-dev-site.
+  * Updated documentation for Piared result graph to be carbon-centric.
+